rename Box.java to AABB.java and remove abstract
Rename the abstract Box.java class to axis-aligned bounding box (AABB).
The name suits the class better, as the Box class was already describing
an AABB.

The AABB is not 'abstract' anymore, as an axis-aligned bounding box is a
'real' object which is used to describe a bounding volume of an object.