implement graph creation
Implement the graph creation. Given an octree, a graph can be created
which represents the spacial relations of the objects in the graph,
namely, that if an edge from node v to node w exists, v must be removed
before v can be removed.
The graph is created by utilizing frustum culling, to determine which
package lay before or above the current object.